+++ pform/pfoo/f_prevprime.cpp
@@ -36,25 +36,25 @@ PFString F_PrevPrime::GetArgumentName(DWORD /*dwIndex*/) const
 
 PFBoolean F_PrevPrime::CallFunction(PFSymbolTable *pContext)
 {
-   PFBoolean bRetval=PFBoolean::b_false;
-   IPFSymbol *pSymbol=pContext->LookupSymbol("_N");
+   PFBoolean bRetval = PFBoolean::b_false;
+   IPFSymbol *pSymbol = pContext->LookupSymbol("_N");
 
    if (!pSymbol) return bRetval;
 
-   if (pSymbol->GetSymbolType()!=INTEGER_SYMBOL_TYPE)  return bRetval;
+   if (pSymbol->GetSymbolType() != INTEGER_SYMBOL_TYPE)  return bRetval;
 
-   Integer *q=((PFIntegerSymbol*)pSymbol)->GetValue();
+   Integer *q = ((PFIntegerSymbol*)pSymbol)->GetValue();
 
    if (!q) return bRetval;
 
-   uint64_t last=(*q)&(ULLONG_MAX);
+   uint64_t last = (*q)&((uint64_t)ULLONG_MAX);
 
-   Integer *r=new Integer;
-   bRetval=PFBoolean::b_true;
+   Integer *r = new Integer;
+   bRetval = PFBoolean::b_true;
 
    *r = primesieve::nth_prime(-1, last);
 
-   pContext->AddSymbol(new PFIntegerSymbol("_result",r));
+   pContext->AddSymbol(new PFIntegerSymbol("_result", r));
 
    return bRetval;
 }
